@@ -1,26 +1,15 @@
-buildscript {
-    repositories {
-        maven { url = "https://maven.minecraftforge.net" }
-        mavenCentral()
-    }
-
-    dependencies {
-        classpath group: "net.minecraftforge.gradle", name: "ForgeGradle", version: "4.1.+", changing: true
-    }
-}
-
 plugins {
     id "com.github.johnrengelman.shadow" version "5.2.0"
     id "eclipse"
     id "idea"
     id "maven-publish"
+    id "net.minecraftforge.gradle"
 }
-apply plugin: "net.minecraftforge.gradle"
 
 version = "$project.minecraft_version-$project.mod_version"
 group = project.maven_group as String // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(8) // Mojang ships Java 8 to end users, so your mod should target Java 8.
+java.toolchain.languageVersion = JavaLanguageVersion.of(16) // Mojang ships Java 16 to end users, so your mod should target Java 16.
 
 println("""
 Project: $project.version
@@ -128,25 +117,34 @@ configurations {
 }
 shadowJar {
     configurations = [project.configurations.shade]
+    relocate "com.tterrag.registrate", "com.crypticcosmos.crypticcosmos.repack.registrate"
 }
 
 build.dependsOn shadowJar
 
 repositories {
-    maven { url("https://repo.repsy.io/mvn/gandiber/geckolib") }
-    maven { // Registrate
-        url "https://maven.tterrag.com/"
-    }
+    maven { url("https://dl.cloudsmith.io/public/geckolib3/geckolib/maven/") }
+    // Registrate
+    maven { url("https://maven.tterrag.com/") }
     mavenLocal()
+    // If you have mod jar dependencies in ./libs, you can declare them as a repository like so:
+    flatDir {
+        dir 'libs'
+    }
+    mavenCentral()
 }
 
 dependencies {
     // Specify the version of Minecraft to use, If this is any group other then "net.minecraft" it is assumed
     // that the dep is a ForgeGradle "patcher" dependency. And it"s patches will be applied.
     minecraft group: "net.minecraftforge", name: "forge", version: "${project.minecraft_version}-${project.forge_version}"
-    implementation fg.deobf("software.bernie.geckolib:geckolib-forge-${project.minecraft_version}:3.+")
+//    implementation fg.deobf("software.bernie.geckolib:geckolib-forge-${project.minecraft_version}:3.+")
+    implementation fg.deobf("software.bernie.geckolib:geckolib-forge-1.16.5:3.+")
 
-    def registrate = "com.tterrag.registrate:Registrate:MC${project.minecraft_version}-${project.registrate_version}"
+//    def registrate = "com.tterrag.registrate:Registrate:MC${project.minecraft_version}-${project.registrate_version}"
+//    def registrate = "com.tterrag.registrate:Registrate:MC1.16.5-${project.registrate_version}"
+    // temporary test version
+    def registrate = "blank:Registrate:MC${project.minecraft_version}-${project.registrate_version}"
     implementation fg.deobf(registrate)
     shade(registrate)
 
