@@ -2,23 +2,21 @@ package com.crypticcosmos.crypticcosmos.block;
 
 import com.crypticcosmos.crypticcosmos.register.EffectRegistries;
 import com.crypticcosmos.crypticcosmos.register.TagRegistries;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.BushBlock;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.potion.EffectInstance;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.shapes.ISelectionContext;
-import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.BushBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
 
 import javax.annotation.Nonnull;
 
 public class MondroveFungus extends BushBlock {
-    public static final VoxelShape SHAPE = Block.box(5.0D,
+    public static final VoxelShape SHAPE = box(5.0D,
             0.0D,
             5.0D,
             11.0D,
@@ -30,15 +28,15 @@ public class MondroveFungus extends BushBlock {
     }
 
     @Override
-    public boolean mayPlaceOn(BlockState state, @Nonnull IBlockReader worldIn, @Nonnull BlockPos pos) {
+    public boolean mayPlaceOn(BlockState state, @Nonnull BlockGetter worldIn, @Nonnull BlockPos pos) {
         return state.is(TagRegistries.MONDROVE_FUNGUS_PLANTABLE);
     }
 
     @SuppressWarnings("deprecation")
     @Override
-    public void entityInside(@Nonnull BlockState state, World world, @Nonnull BlockPos pos, @Nonnull Entity entity) {
-        if (!world.isClientSide() && entity instanceof LivingEntity) {
-            ((LivingEntity) entity).addEffect(new EffectInstance(EffectRegistries.CORRUPTION.get(), Integer.MAX_VALUE));
+    public void entityInside(@Nonnull BlockState state, Level world, @Nonnull BlockPos pos, @Nonnull Entity entity) {
+        if (!world.isClientSide() && entity instanceof LivingEntity livingEntity) {
+            livingEntity.addEffect(new MobEffectInstance(EffectRegistries.CORRUPTION.get(), Integer.MAX_VALUE));
         }
     }
 
@@ -46,10 +44,10 @@ public class MondroveFungus extends BushBlock {
     @Override
     @Nonnull
     public VoxelShape getShape(BlockState state,
-                               @Nonnull IBlockReader worldIn,
+                               @Nonnull BlockGetter worldIn,
                                @Nonnull BlockPos pos,
-                               @Nonnull ISelectionContext context) {
-        Vector3d vec3d = state.getOffset(worldIn, pos);
+                               @Nonnull CollisionContext context) {
+        var vec3d = state.getOffset(worldIn, pos);
         return SHAPE.move(vec3d.x, vec3d.y, vec3d.z);
     }
 }
