@@ -10,12 +10,12 @@ import com.crypticcosmos.crypticcosmos.world.feature.ConfiguredFeatureRegistries
 import com.crypticcosmos.crypticcosmos.world.feature.FeatureRegistries;
 import com.crypticcosmos.crypticcosmos.world.structures.StructureConfig;
 import com.tterrag.registrate.Registrate;
-import net.minecraft.block.WoodType;
-import net.minecraft.client.renderer.Atlases;
+import net.minecraft.client.renderer.Sheets;
 import net.minecraft.data.DataGenerator;
-import net.minecraft.item.ItemGroup;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.state.properties.WoodType;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.common.util.Lazy;
@@ -24,8 +24,8 @@ import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.GatherDataEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.forge.event.lifecycle.GatherDataEvent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import software.bernie.geckolib3.GeckoLib;
@@ -38,7 +38,7 @@ import static com.crypticcosmos.crypticcosmos.register.SignRegistries.*;
 public class CrypticCosmos {
     public static final Logger LOGGER = LogManager.getLogger(CrypticCosmos.class);
     public static final String MOD_ID = "crypticcosmos";
-    public static final ItemGroup CRYPTIC_COSMOS_ITEM_GROUP = new ItemGroup("cryptic_cosmos_tab") {
+    public static final CreativeModeTab CRYPTIC_COSMOS_ITEM_GROUP = new CreativeModeTab("cryptic_cosmos_tab") {
         @Nonnull
         @Override
         public ItemStack makeIcon() {
@@ -99,9 +99,9 @@ public class CrypticCosmos {
 
     private void clientSetup(final FMLClientSetupEvent event) {
         event.enqueueWork(() -> {
-            Atlases.addWoodType(MONDROVE_WOOD_TYPE);
-            Atlases.addWoodType(GROMBLE_WOOD_TYPE);
-            Atlases.addWoodType(OSMINSTEM_WOOD_TYPE);
+            Sheets.addWoodType(MONDROVE_WOOD_TYPE);
+            Sheets.addWoodType(GROMBLE_WOOD_TYPE);
+            Sheets.addWoodType(OSMINSTEM_WOOD_TYPE);
         });
     }
 
