@@ -1,24 +1,22 @@
 package com.crypticcosmos.crypticcosmos.block;
 
 import com.crypticcosmos.crypticcosmos.register.BlockRegistries;
-import net.minecraft.block.AbstractBlock;
-import net.minecraft.block.AbstractBodyPlantBlock;
-import net.minecraft.block.AbstractTopPlantBlock;
-import net.minecraft.block.Block;
-import net.minecraft.util.Direction;
-import net.minecraft.util.math.shapes.VoxelShape;
+import net.minecraft.core.Direction;
+import net.minecraft.world.level.block.GrowingPlantBodyBlock;
+import net.minecraft.world.level.block.GrowingPlantHeadBlock;
+import net.minecraft.world.phys.shapes.VoxelShape;
 
 import javax.annotation.Nonnull;
 
-public class ArtificialGrombleStalk extends AbstractBodyPlantBlock {
-    public static final VoxelShape SHAPE = Block.box(4.0D, 0.0D, 4.0D, 12.0D, 16.0D, 12.0D);
+public class ArtificialGrombleStalk extends GrowingPlantBodyBlock {
+    public static final VoxelShape SHAPE = box(4.0D, 0.0D, 4.0D, 12.0D, 16.0D, 12.0D);
 
-    public ArtificialGrombleStalk(AbstractBlock.Properties p) {
+    public ArtificialGrombleStalk(Properties p) {
         super(p, Direction.UP, SHAPE, false);
     }
 
     @Nonnull
-    public AbstractTopPlantBlock getHeadBlock() {
+    public GrowingPlantHeadBlock getHeadBlock() {
         return BlockRegistries.ARTIFICIAL_GROMBLE_STALK.get();
     }
 }
