@@ -1,17 +1,16 @@
 package com.crypticcosmos.crypticcosmos.block;
 
 import com.crypticcosmos.crypticcosmos.register.CrypticCosmosDimensions;
-import net.minecraft.block.BlockState;
-import net.minecraft.state.IntegerProperty;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.server.ServerWorld;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
 
 public interface Infectable {
     IntegerProperty INFECTION_LEVEL = IntegerProperty.create("infection_level", 0, 2);
 
-    default void infect(ServerWorld world, BlockPos pos) {
+    default void infect(ServerLevel world, BlockPos pos) {
         if (world.dimension() == CrypticCosmosDimensions.MAKROSSA_KEY) {
-            BlockState infectedLogState = world.getBlockState(pos);
+            var infectedLogState = world.getBlockState(pos);
 
             if (infectedLogState.getBlock() instanceof Infectable &&
                 infectedLogState.getValue(INFECTION_LEVEL) < 2) {
