@@ -2,15 +2,15 @@ package com.crypticcosmos.crypticcosmos.block;
 
 import com.crypticcosmos.crypticcosmos.util.Strippable;
 import com.tterrag.registrate.util.entry.BlockEntry;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.RotatedPillarBlock;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.ActionResultType;
-import net.minecraft.util.Hand;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.BlockRayTraceResult;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.RotatedPillarBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
 
 import javax.annotation.Nonnull;
 
@@ -32,12 +32,12 @@ public class OsminstemLog extends RotatedPillarBlock implements Strippable {
     @SuppressWarnings("deprecation")
     @Nonnull
     @Override
-    public ActionResultType use(@Nonnull BlockState state,
-                                @Nonnull World world,
-                                @Nonnull BlockPos pos,
-                                @Nonnull PlayerEntity player,
-                                @Nonnull Hand hand,
-                                @Nonnull BlockRayTraceResult rayTraceResult) {
+    public InteractionResult use(@Nonnull BlockState state,
+                                 @Nonnull Level world,
+                                 @Nonnull BlockPos pos,
+                                 @Nonnull Player player,
+                                 @Nonnull InteractionHand hand,
+                                 @Nonnull BlockHitResult rayTraceResult) {
         return strip(strippedBlock,
                 player.getItemInHand(hand),
                 world, pos, state, player, hand);
