@@ -1,23 +1,23 @@
 package com.crypticcosmos.crypticcosmos.entity.creature.gromble_snatcher;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
-import com.mojang.blaze3d.vertex.IVertexBuilder;
-import net.minecraft.client.renderer.IRenderTypeBuffer;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.entity.EntityRendererManager;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.resources.ResourceLocation;
 import software.bernie.geckolib3.renderers.geo.GeoEntityRenderer;
 
 import javax.annotation.Nullable;
 
-public class GrombleSnatcherRender extends GeoEntityRenderer<GrombleSnatcherEntity> {
-    public GrombleSnatcherRender(EntityRendererManager renderManager) {
-        super(renderManager, new GrombleSnatcherModel());
+public class GrombleSnatcherRender extends GeoEntityRenderer<GrombleSnatcher> {
+    public GrombleSnatcherRender(EntityRendererProvider.Context context) {
+        super(context, new GrombleSnatcherModel());
         this.shadowRadius = 1.8F;
     }
 
     @Override
-    public RenderType getRenderType(GrombleSnatcherEntity animatable, float partialTicks, MatrixStack stack, @Nullable IRenderTypeBuffer renderTypeBuffer, @Nullable IVertexBuilder vertexBuilder, int packedLightIn, ResourceLocation textureLocation) {
+    public RenderType getRenderType(GrombleSnatcher animatable, float partialTicks, PoseStack stack, @Nullable MultiBufferSource renderTypeBuffer, @Nullable VertexConsumer vertexBuilder, int packedLightIn, ResourceLocation textureLocation) {
         return RenderType.entityCutoutNoCull(textureLocation);
     }
 }
