@@ -2,13 +2,13 @@ package com.crypticcosmos.crypticcosmos.register;
 
 import com.crypticcosmos.crypticcosmos.CrypticCosmos;
 import com.crypticcosmos.crypticcosmos.world.biomes.BiomeMaker;
-import net.minecraft.util.RegistryKey;
-import net.minecraft.util.registry.Registry;
-import net.minecraft.world.biome.Biome;
+import net.minecraft.core.Registry;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.level.biome.Biome;
 import net.minecraftforge.common.BiomeDictionary;
 import net.minecraftforge.common.BiomeDictionary.Type;
-import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -17,24 +17,24 @@ public class BiomeRegistries {
 
     // lunara biomes
     public static final RegistryObject<Biome> LUNARAN_PLAINS = BIOMES.register("lunaran_plains", BiomeMaker::lunaraPlains);
-    public static final RegistryKey<Biome> LUNARAN_PLAINS_KEY =
-            RegistryKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("lunaran_plains"));
+    public static final ResourceKey<Biome> LUNARAN_PLAINS_KEY =
+            ResourceKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("lunaran_plains"));
 
     public static final RegistryObject<Biome> ACERBIC_ISLES = BIOMES.register("acerbic_isles", BiomeMaker::acerbicIsles);
-    public static final RegistryKey<Biome> ACERBIC_ISLES_KEY =
-            RegistryKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("acerbic_isles"));
+    public static final ResourceKey<Biome> ACERBIC_ISLES_KEY =
+            ResourceKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("acerbic_isles"));
 
     public static final RegistryObject<Biome> KAFSINIAN_FOREST = BIOMES.register("kafsinian_forest", BiomeMaker::kafisnianForest);
-    public static final RegistryKey<Biome> KAFSINIAN_FOREST_KEY =
-            RegistryKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("kafsinian_forest"));
+    public static final ResourceKey<Biome> KAFSINIAN_FOREST_KEY =
+            ResourceKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("kafsinian_forest"));
 
     public static final RegistryObject<Biome> GROMBLE_GROVE = BIOMES.register("gromble_grove", BiomeMaker::grombleGrove);
-    public static final RegistryKey<Biome> GROMBLE_GROVE_KEY =
-            RegistryKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("gromble_grove"));
+    public static final ResourceKey<Biome> GROMBLE_GROVE_KEY =
+            ResourceKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("gromble_grove"));
 
     public static final RegistryObject<Biome> UMBRAL_DUNES = BIOMES.register("umbral_dunes", BiomeMaker::umbralDunes);
-    public static final RegistryKey<Biome> UMBRAL_DUNES_KEY =
-            RegistryKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("umbral_dunes"));
+    public static final ResourceKey<Biome> UMBRAL_DUNES_KEY =
+            ResourceKey.create(Registry.BIOME_REGISTRY, CrypticCosmos.id("umbral_dunes"));
 
     public static void biomeLoading(FMLCommonSetupEvent event) {
         BiomeDictionary.addTypes(LUNARAN_PLAINS_KEY, Type.PLAINS, Type.DRY, BiomeDictionary.Type.COLD, Type.DEAD);
