@@ -2,30 +2,26 @@ package com.crypticcosmos.crypticcosmos.register;
 
 import com.crypticcosmos.crypticcosmos.CrypticCosmos;
 import com.crypticcosmos.crypticcosmos.world.structures.MondroveBundleStructure;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import net.minecraft.util.registry.WorldGenRegistries;
-import net.minecraft.world.gen.feature.NoFeatureConfig;
-import net.minecraft.world.gen.feature.structure.Structure;
-import net.minecraft.world.gen.settings.DimensionStructuresSettings;
-import net.minecraft.world.gen.settings.StructureSeparationSettings;
-import net.minecraftforge.fml.RegistryObject;
+import net.minecraft.data.BuiltinRegistries;
+import net.minecraft.world.level.levelgen.StructureSettings;
+import net.minecraft.world.level.levelgen.feature.StructureFeature;
+import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.StructureFeatureConfiguration;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
-import java.util.HashMap;
-import java.util.Map;
-
 public class StructureRegistries {
-    public static final DeferredRegister<Structure<?>> DEFERRED_REGISTRY_STRUCTURE = DeferredRegister.create(ForgeRegistries.STRUCTURE_FEATURES, CrypticCosmos.MOD_ID);
+    public static final DeferredRegister<StructureFeature<?>> DEFERRED_REGISTRY_STRUCTURE = DeferredRegister.create(ForgeRegistries.STRUCTURE_FEATURES, CrypticCosmos.MOD_ID);
 
 
-    public static final RegistryObject<Structure<NoFeatureConfig>> MONDROVE_BUNDLE = DEFERRED_REGISTRY_STRUCTURE.register("mondrove_bundle", () -> new MondroveBundleStructure(NoFeatureConfig.CODEC));
+    public static final RegistryObject<StructureFeature<NoneFeatureConfiguration>> MONDROVE_BUNDLE = DEFERRED_REGISTRY_STRUCTURE.register("mondrove_bundle", () -> new MondroveBundleStructure(NoneFeatureConfiguration.CODEC));
 
     public static void setupStructures() {
         setupMapSpacingAndLand(
                 MONDROVE_BUNDLE.get(), /* The instance of the structure */
-                new StructureSeparationSettings(10 /* average distance apart in chunks between spawn attempts */,
+                new StructureFeatureConfiguration(10 /* average distance apart in chunks between spawn attempts */,
                         5 /* minimum distance apart in chunks between spawn attempts */,
                         439096529 /* this modifies the seed of the structure so no two structures always spawn over each-other. Make this large and unique. */),
                 true);
@@ -34,9 +30,9 @@ public class StructureRegistries {
         // Add more structures here and so on
     }
 
-    public static <F extends Structure<?>> void setupMapSpacingAndLand(
+    public static <F extends StructureFeature<?>> void setupMapSpacingAndLand(
             F structure,
-            StructureSeparationSettings structureSeparationSettings,
+            StructureFeatureConfiguration structureFeatureConfig,
             boolean transformSurroundingLand) { /*
      * We need to add our structures into the map in Structure class
      * alongside vanilla structures or else it will cause errors.
@@ -45,7 +41,7 @@ public class StructureRegistries {
      * getRegistryName() should never return null.
      */
         //noinspection ConstantConditions
-        Structure.STRUCTURES_REGISTRY.put(structure.getRegistryName().toString(), structure);
+        StructureFeature.STRUCTURES_REGISTRY.put(structure.getRegistryName().toString(), structure);
 
         /*
          * Whether surrounding land will be modified automatically to conform to the bottom of the structure.
@@ -57,13 +53,7 @@ public class StructureRegistries {
          *
          * NOISE_AFFECTING_FEATURES requires AccessTransformer  (See resources/META-INF/accesstransformer.cfg)
          */
-        if (transformSurroundingLand) {
-            Structure.NOISE_AFFECTING_FEATURES =
-                    ImmutableList.<Structure<?>>builder()
-                            .addAll(Structure.NOISE_AFFECTING_FEATURES)
-                            .add(structure)
-                            .build();
-        }
+        if (transformSurroundingLand) StructureFeature.NOISE_AFFECTING_FEATURES.add(structure);
         /*
          * This is the map that holds the default spacing of all structures.
          * Always add your structure to here so that other mods can utilize it if needed.
@@ -77,12 +67,7 @@ public class StructureRegistries {
          *
          * DEFAULTS requires AccessTransformer  (See resources/META-INF/accesstransformer.cfg)
          */
-        DimensionStructuresSettings.DEFAULTS =
-                ImmutableMap.<Structure<?>, StructureSeparationSettings>builder()
-                        .putAll(DimensionStructuresSettings.DEFAULTS)
-                        .put(structure, structureSeparationSettings)
-                        .build();
-
+        StructureSettings.DEFAULTS.put(structure, structureFeatureConfig);
 
         /*
          * There are very few mods that relies on seeing your structure in the noise settings registry before the world is made.
@@ -91,8 +76,8 @@ public class StructureRegistries {
          * that field only applies for the default overworld and won't add to other worldtypes or dimensions (like amplified or Nether).
          * So yeah, don't do DimensionSettings.BUILTIN_OVERWORLD. Use the NOISE_GENERATOR_SETTINGS loop below instead if you must.
          */
-        WorldGenRegistries.NOISE_GENERATOR_SETTINGS.entrySet().forEach(settings -> {
-            Map<Structure<?>, StructureSeparationSettings> structureMap = settings.getValue().structureSettings().structureConfig();
+        BuiltinRegistries.NOISE_GENERATOR_SETTINGS.entrySet().forEach(settings -> {
+            var structureMap = settings.getValue().structureSettings().structureConfig();
 
             /*
              * Pre-caution in case a mod makes the structure map immutable like datapacks do.
@@ -100,14 +85,10 @@ public class StructureRegistries {
              *
              * structureConfig requires AccessTransformer  (See resources/META-INF/accesstransformer.cfg)
              */
-            if(structureMap instanceof ImmutableMap){
-                Map<Structure<?>, StructureSeparationSettings> tempMap = new HashMap<>(structureMap);
-                tempMap.put(structure, structureSeparationSettings);
-                settings.getValue().structureSettings().structureConfig = tempMap;
-            }
-            else{
-                structureMap.put(structure, structureSeparationSettings);
-            }
+            if (structureMap instanceof ImmutableMap<StructureFeature<?>, StructureFeatureConfiguration> tempMap) {
+                settings.getValue().structureSettings().structureConfig().clear();
+                settings.getValue().structureSettings().structureConfig().putAll(tempMap);
+            } else structureMap.put(structure, structureFeatureConfig);
         });
     }
 }
