@@ -9,37 +9,33 @@ import com.tterrag.registrate.providers.DataGenContext;
 import com.tterrag.registrate.providers.RegistrateBlockstateProvider;
 import com.tterrag.registrate.providers.loot.RegistrateBlockLootTables;
 import com.tterrag.registrate.util.nullness.NonNullSupplier;
-import net.minecraft.advancements.criterion.EnchantmentPredicate;
-import net.minecraft.block.*;
-import net.minecraft.enchantment.Enchantments;
-import net.minecraft.item.Item;
-import net.minecraft.loot.LootTable;
-import net.minecraft.state.properties.AttachFace;
-import net.minecraft.state.properties.BlockStateProperties;
-import net.minecraft.util.Direction;
-import net.minecraft.util.Direction.Axis;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.advancements.critereon.EnchantmentPredicate;
+import net.minecraft.core.Direction.Axis;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.*;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraftforge.client.model.generators.ConfiguredModel;
 
 import java.util.function.Supplier;
 
-import static com.crypticcosmos.crypticcosmos.CrypticCosmos.id;
-import static net.minecraft.advancements.criterion.ItemPredicate.Builder.item;
-import static net.minecraft.advancements.criterion.MinMaxBounds.IntBound.atLeast;
-import static net.minecraft.enchantment.Enchantments.BLOCK_FORTUNE;
-import static net.minecraft.enchantment.Enchantments.SILK_TOUCH;
-import static net.minecraft.loot.AlternativesLootEntry.alternatives;
-import static net.minecraft.loot.ConstantRange.exactly;
-import static net.minecraft.loot.ItemLootEntry.lootTableItem;
-import static net.minecraft.loot.LootPool.lootPool;
-import static net.minecraft.loot.LootTable.lootTable;
-import static net.minecraft.loot.RandomValueRange.between;
-import static net.minecraft.loot.conditions.Alternative.alternative;
-import static net.minecraft.loot.conditions.MatchTool.toolMatches;
-import static net.minecraft.loot.conditions.TableBonus.bonusLevelFlatChance;
-import static net.minecraft.loot.functions.ApplyBonus.addOreBonusCount;
-import static net.minecraft.loot.functions.ExplosionDecay.explosionDecay;
-import static net.minecraft.loot.functions.SetCount.setCount;
+import static net.minecraft.advancements.critereon.ItemPredicate.Builder.item;
+import static net.minecraft.advancements.critereon.MinMaxBounds.Ints.atLeast;
+import static net.minecraft.world.item.enchantment.Enchantments.BLOCK_FORTUNE;
+import static net.minecraft.world.item.enchantment.Enchantments.SILK_TOUCH;
+import static net.minecraft.world.level.storage.loot.LootPool.lootPool;
+import static net.minecraft.world.level.storage.loot.LootTable.lootTable;
+import static net.minecraft.world.level.storage.loot.entries.AlternativesEntry.alternatives;
+import static net.minecraft.world.level.storage.loot.entries.LootItem.lootTableItem;
+import static net.minecraft.world.level.storage.loot.functions.ApplyBonusCount.addOreBonusCount;
+import static net.minecraft.world.level.storage.loot.functions.ApplyExplosionDecay.explosionDecay;
+import static net.minecraft.world.level.storage.loot.functions.SetItemCountFunction.setCount;
+import static net.minecraft.world.level.storage.loot.predicates.AlternativeLootItemCondition.alternative;
+import static net.minecraft.world.level.storage.loot.predicates.BonusLevelTableCondition.bonusLevelFlatChance;
+import static net.minecraft.world.level.storage.loot.predicates.MatchTool.toolMatches;
+import static net.minecraft.world.level.storage.loot.providers.number.ConstantValue.exactly;
+import static net.minecraft.world.level.storage.loot.providers.number.UniformGenerator.between;
 import static net.minecraftforge.client.model.generators.ConfiguredModel.builder;
 import static net.minecraftforge.common.Tags.Items.SHEARS;
 
@@ -52,9 +48,9 @@ public class RegistrationUtils {
 
         return builder()
                 .modelFile(provider.models().cubeAll(context.getName() + infectionLevel,
-                        blockTexture(context, infectionLevel)
-                ))
-                .build();
+                                blockTexture(context, infectionLevel)
+                        )
+                ).build();
     }
 
     public static ConfiguredModel[] mondroveLogModels(BlockState state,
@@ -63,12 +59,12 @@ public class RegistrationUtils {
                                                       boolean isStripped) {
         String infectionLevel = state.getValue(Infectable.INFECTION_LEVEL) == 0 ? ""
                 : state.getValue(Infectable.INFECTION_LEVEL).toString();
-        Axis axis = getAxisFromBlockState(state);
+        var axis = state.getValue(BlockStateProperties.AXIS);
 
         return builder().modelFile(provider.models().cubeColumn(
-                context.getName() + infectionLevel,
-                blockTexture(isStripped ? "stripped_" : "", MondroveRegistries.MONDROVE_LOG.get(), !isStripped ? infectionLevel : ""),
-                blockTexture(isStripped ? "stripped_" : "", MondroveRegistries.MONDROVE_LOG.get(), "_top")))
+                        context.getName() + infectionLevel,
+                        blockTexture(isStripped ? "stripped_" : "", MondroveRegistries.MONDROVE_LOG, !isStripped ? infectionLevel : ""),
+                        blockTexture(isStripped ? "stripped_" : "", MondroveRegistries.MONDROVE_LOG, "_top")))
                 .rotationX(axis.equals(Axis.Z) || axis.equals(Axis.X) ? 90 : 0)
                 .rotationY(axis.equals(Axis.X) ? 90 : 0).build();
     }
@@ -77,17 +73,17 @@ public class RegistrationUtils {
                                                        DataGenContext<Block, MondroveLog> context,
                                                        RegistrateBlockstateProvider provider,
                                                        boolean isStripped) {
-        Axis axis = getAxisFromBlockState(state);
+        var axis = state.getValue(BlockStateProperties.AXIS);
         String infectionLevel = state.getValue(Infectable.INFECTION_LEVEL) == 0 ? ""
                 : state.getValue(Infectable.INFECTION_LEVEL).toString();
 
         return builder().modelFile(provider.models().cubeAll(
-                context.getName() + infectionLevel,
-                blockTexture(isStripped ? "stripped_" : "",
-                        MondroveRegistries.MONDROVE_LOG.get(),
-                        !isStripped ? infectionLevel : ""
-                ))
-        ).rotationX(axis.equals(Axis.Z) || axis.equals(Axis.X) ? 90 : 0)
+                        context.getName() + infectionLevel,
+                        blockTexture(isStripped ? "stripped_" : "",
+                                MondroveRegistries.MONDROVE_LOG,
+                                !isStripped ? infectionLevel : ""
+                        ))
+                ).rotationX(axis.equals(Axis.Z) || axis.equals(Axis.X) ? 90 : 0)
                 .rotationY(axis.equals(Axis.X) ? 90 : 0).build();
     }
 
@@ -95,17 +91,17 @@ public class RegistrationUtils {
         provider.getVariantBuilder(context.get()).forAllStates(
                 blockState -> {
                     boolean isSnowy = blockState.getValue(OvergrownLunonBlock.SNOWY);
-                    Direction axis = blockState.getValue(OvergrownLunonBlock.FACING);
+                    var axis = blockState.getValue(OvergrownLunonBlock.FACING);
                     final ConfiguredModel.Builder<?> builder = builder();
 
                     int degree = (int) axis.toYRot();
 
                     if (!isSnowy) {
                         builder.modelFile(provider.models().cubeBottomTop(
-                                context.getName(),
-                                blockTexture(context, "_side"),
-                                provider.blockTexture(LunonRegistries.LUNON.get()),
-                                blockTexture(context, "_top")))
+                                        context.getName(),
+                                        blockTexture(context, "_side"),
+                                        provider.blockTexture(LunonRegistries.LUNON.get()),
+                                        blockTexture(context, "_top")))
                                 .rotationY(degree);
                     } else {
                         builder.modelFile(provider.models().cubeBottomTop(
@@ -123,9 +119,9 @@ public class RegistrationUtils {
     public static void leavesModel(DataGenContext<Block, LeavesBlock> context, RegistrateBlockstateProvider provider) {
         provider.getVariantBuilder(context.get())
                 .forAllStates(state -> builder().modelFile(
-                        provider.models()
-                                .withExistingParent(context.getName(), provider.mcLoc("leaves"))
-                                .texture("all", provider.blockTexture(context.get()))
+                                provider.models()
+                                        .withExistingParent(context.getName(), provider.mcLoc("leaves"))
+                                        .texture("all", provider.blockTexture(context.get()))
                         ).build()
                 );
     }
@@ -143,7 +139,7 @@ public class RegistrationUtils {
                 RegistrationUtils.blockTexture(context, "_top"));
     }
 
-    public static void buttonModel(DataGenContext<Block, ? extends AbstractButtonBlock> context, RegistrateBlockstateProvider provider, NonNullSupplier<? extends Block> planks) {
+    public static void buttonModel(DataGenContext<Block, ? extends ButtonBlock> context, RegistrateBlockstateProvider provider, NonNullSupplier<? extends Block> planks) {
         final ConfiguredModel.Builder<?> buttonModel = builder().modelFile(provider.models()
                 .withExistingParent(context.getName(), "button")
                 .texture("texture", provider.blockTexture(planks.get())));
@@ -159,18 +155,22 @@ public class RegistrationUtils {
 
         provider.getVariantBuilder(context.get())
                 .forAllStates(state -> {
-                    final AttachFace attachedFace = state.getValue(AbstractButtonBlock.FACE);
-                    final Direction direction = state.getValue(AbstractButtonBlock.FACING);
-                    final boolean isPowered = state.getValue(AbstractButtonBlock.POWERED);
+                    final var attachedFace = state.getValue(ButtonBlock.FACE);
+                    final var direction = state.getValue(ButtonBlock.FACING);
+                    final boolean isPowered = state.getValue(ButtonBlock.POWERED);
 
-                    int rotationY = 0;
-                    if (direction.equals(Direction.EAST)) rotationY = 90;
-                    if (direction.equals(Direction.SOUTH)) rotationY = 180;
-                    if (direction.equals(Direction.WEST)) rotationY = 270;
+                    int rotationY = switch (direction) {
+                        case NORTH -> 0;
+                        case EAST -> 90;
+                        case SOUTH -> 180;
+                        default -> 270;
+                    };
 
-                    int rotationX = 0;
-                    if (attachedFace.equals(AttachFace.CEILING)) rotationX = 180;
-                    else if (attachedFace.equals(AttachFace.WALL)) rotationX = 90;
+                    int rotationX = switch (attachedFace) {
+                        case FLOOR -> 0;
+                        case WALL -> 90;
+                        case CEILING -> 180;
+                    };
 
                     if (!isPowered) return buttonModel.rotationX(rotationX).rotationY(rotationY).build();
                     else return buttonPressedModel.rotationX(rotationX).rotationY(rotationY).build();
@@ -191,18 +191,18 @@ public class RegistrationUtils {
                 });
     }
 
-    public static Axis getAxisFromBlockState(BlockState state) {
-        return state.getValue(BlockStateProperties.AXIS);
-    }
-
     public static ResourceLocation blockTexture(Supplier<? extends Block> block, String suffix) {
         //noinspection ConstantConditions
-        return id("block/" + block.get().getRegistryName().getPath() + suffix);
+        return new ResourceLocation(block.get().getRegistryName().getNamespace(),
+                "block/" + block.get().getRegistryName().getPath() + suffix
+        );
     }
 
-    public static ResourceLocation blockTexture(String prefix, Block block, String suffix) {
+    public static ResourceLocation blockTexture(String prefix, Supplier<? extends Block> block, String suffix) {
         //noinspection ConstantConditions
-        return id("block/" + prefix + block.getRegistryName().getPath() + suffix);
+        return new ResourceLocation(block.get().getRegistryName().getNamespace(),
+                "block/" + prefix + block.get().getRegistryName().getPath() + suffix
+        );
     }
 
     public static <BLOCK extends Block, ITEM extends NonNullSupplier<Item>> void silkTouchFortune(
@@ -227,7 +227,7 @@ public class RegistrationUtils {
                                                 ? between(minDrop, maxDrop)
                                                 : exactly(minDrop)
                                         ))
-                                        .apply(addOreBonusCount(Enchantments.BLOCK_FORTUNE))
+                                        .apply(addOreBonusCount(BLOCK_FORTUNE))
                                         .apply(explosionDecay())
                         ))
                 )
@@ -235,7 +235,7 @@ public class RegistrationUtils {
     }
 
     public static <TOP extends Block, BODY extends Block> void vinesLootTable(RegistrateBlockLootTables lootTables, TOP topVine, BODY bodyVine) {
-        final LootTable.Builder lootTable = lootTable()
+        final var lootTable = lootTable()
                 .withPool(lootPool()
                         .add(lootTableItem(topVine)
                                 .when(alternative(
