@@ -4,103 +4,102 @@ import com.crypticcosmos.crypticcosmos.CrypticCosmos;
 import com.crypticcosmos.crypticcosmos.register.EntityTypeRegistries;
 import com.google.common.collect.ImmutableList;
 import com.mojang.serialization.Codec;
-import net.minecraft.block.BlockState;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.SharedSeedRandom;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.ChunkPos;
-import net.minecraft.util.math.MutableBoundingBox;
-import net.minecraft.util.registry.DynamicRegistries;
-import net.minecraft.util.registry.Registry;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.biome.Biome;
-import net.minecraft.world.biome.MobSpawnInfo;
-import net.minecraft.world.biome.provider.BiomeProvider;
-import net.minecraft.world.gen.ChunkGenerator;
-import net.minecraft.world.gen.GenerationStage;
-import net.minecraft.world.gen.Heightmap;
-import net.minecraft.world.gen.feature.NoFeatureConfig;
-import net.minecraft.world.gen.feature.jigsaw.JigsawManager;
-import net.minecraft.world.gen.feature.structure.AbstractVillagePiece;
-import net.minecraft.world.gen.feature.structure.Structure;
-import net.minecraft.world.gen.feature.structure.StructureStart;
-import net.minecraft.world.gen.feature.structure.VillageConfig;
-import net.minecraft.world.gen.feature.template.TemplateManager;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Registry;
+import net.minecraft.core.RegistryAccess;
+import net.minecraft.world.level.ChunkPos;
+import net.minecraft.world.level.LevelHeightAccessor;
+import net.minecraft.world.level.biome.Biome;
+import net.minecraft.world.level.biome.BiomeSource;
+import net.minecraft.world.level.biome.MobSpawnSettings;
+import net.minecraft.world.level.chunk.ChunkGenerator;
+import net.minecraft.world.level.levelgen.GenerationStep;
+import net.minecraft.world.level.levelgen.Heightmap;
+import net.minecraft.world.level.levelgen.WorldgenRandom;
+import net.minecraft.world.level.levelgen.feature.StructureFeature;
+import net.minecraft.world.level.levelgen.feature.configurations.JigsawConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.structures.JigsawPlacement;
+import net.minecraft.world.level.levelgen.structure.PoolElementStructurePiece;
+import net.minecraft.world.level.levelgen.structure.StructureStart;
+import net.minecraft.world.level.levelgen.structure.templatesystem.StructureManager;
 import org.apache.logging.log4j.Level;
 
 import javax.annotation.Nonnull;
 import java.util.List;
 
-public class MondroveBundleStructure extends Structure<NoFeatureConfig> {
+import static com.crypticcosmos.crypticcosmos.CrypticCosmos.id;
 
-    public MondroveBundleStructure(Codec<NoFeatureConfig> codec) {
+public class MondroveBundleStructure extends StructureFeature<NoneFeatureConfiguration> {
+
+    public MondroveBundleStructure(Codec<NoneFeatureConfiguration> codec) {
         super(codec);
     }
 
     @Nonnull
     @Override
-    public IStartFactory<NoFeatureConfig> getStartFactory() {
+    public StructureStartFactory<NoneFeatureConfiguration> getStartFactory() {
         return MondroveBundleStructure.Start::new;
     }
 
     @Nonnull
     @Override
-    public GenerationStage.Decoration step() {
-        return GenerationStage.Decoration.SURFACE_STRUCTURES;
+    public GenerationStep.Decoration step() {
+        return GenerationStep.Decoration.SURFACE_STRUCTURES;
     }
 
     @Override
-    public List<MobSpawnInfo.Spawners> getDefaultCreatureSpawnList() {
+    public List<MobSpawnSettings.SpawnerData> getDefaultCreatureSpawnList() {
         return ImmutableList.of(
-                new MobSpawnInfo.Spawners(EntityTypeRegistries.GROMBLE_FROG.get(), 100, 4, 9)
+                new MobSpawnSettings.SpawnerData(EntityTypeRegistries.GROMBLE_FROG.get(), 100, 4, 9)
         );
     }
 
     @Override
     protected boolean isFeatureChunk(ChunkGenerator chunkGenerator,
-                                     @Nonnull BiomeProvider biomeSource,
+                                     @Nonnull BiomeSource biomeSource,
                                      long seed,
-                                     @Nonnull SharedSeedRandom chunkRandom,
-                                     int chunkX,
-                                     int chunkZ,
-                                     @Nonnull Biome biome,
+                                     @Nonnull WorldgenRandom chunkRandom,
                                      @Nonnull ChunkPos chunkPos,
-                                     @Nonnull NoFeatureConfig featureConfig) {
-        BlockPos centerOfChunk = new BlockPos((chunkX << 4) + 7, 0, (chunkZ << 4) + 7);
+                                     @Nonnull Biome biome,
+                                     @Nonnull ChunkPos chunkPosA,
+                                     @Nonnull NoneFeatureConfiguration featureConfig,
+                                     LevelHeightAccessor accessor) {
+        var centerOfChunk = new BlockPos((chunkPos.x << 4) + 7, 0, (chunkPos.z << 4) + 7);
 
         // Grab height of land. Will stop at first non-air block.
-        int landHeight = chunkGenerator.getFirstOccupiedHeight(centerOfChunk.getX(), centerOfChunk.getZ(), Heightmap.Type.WORLD_SURFACE_WG);
+        int landHeight = chunkGenerator.getFirstOccupiedHeight(centerOfChunk.getX(), centerOfChunk.getZ(), Heightmap.Types.WORLD_SURFACE_WG, accessor);
 
         // Grabs column of blocks at given position. In the overworld, this column will be made of stone, water, and air.
         // In nether, it will be netherrack, lava, and air. End will only be end stone and air. It depends on what block
         // the chunk generator will place for that dimension.
-        IBlockReader columnOfBlocks = chunkGenerator.getBaseColumn(centerOfChunk.getX(), centerOfChunk.getZ());
+        var columnOfBlocks = chunkGenerator.getBaseColumn(centerOfChunk.getX(), centerOfChunk.getZ(), accessor);
 
         // Combine the column of blocks with land height and you get the top block itself which you can test.
-        BlockState topBlock = columnOfBlocks.getBlockState(centerOfChunk.above(landHeight));
+        var topBlock = columnOfBlocks.getBlockState(centerOfChunk.above(landHeight));
 
         // Now we test to make sure our structure is not spawning on water or other fluids.
         // You can do height check instead too to make it spawn at high elevations.
         return topBlock.getFluidState().isEmpty(); //landHeight > 100;
     }
 
-    public static class Start extends StructureStart<NoFeatureConfig> {
-        public Start(Structure<NoFeatureConfig> structureIn, int chunkX, int chunkZ, MutableBoundingBox mutableBoundingBox, int referenceIn, long seedIn) {
-            super(structureIn, chunkX, chunkZ, mutableBoundingBox, referenceIn, seedIn);
+    public static class Start extends StructureStart<NoneFeatureConfiguration> {
+        public Start(StructureFeature<NoneFeatureConfiguration> structure, ChunkPos chunkPos, int reference, long seed) {
+            super(structure, chunkPos, reference, seed);
         }
 
         @Override
-        public void generatePieces(@Nonnull DynamicRegistries dynamicRegistryManager,
+        public void generatePieces(@Nonnull RegistryAccess dynamicRegistryManager,
                                    @Nonnull ChunkGenerator chunkGenerator,
-                                   @Nonnull TemplateManager templateManagerIn,
-                                   int chunkX,
-                                   int chunkZ,
+                                   @Nonnull StructureManager structureManager,
+                                   @Nonnull ChunkPos chunkPos,
                                    @Nonnull Biome biomeIn,
-                                   @Nonnull NoFeatureConfig config) {
+                                   @Nonnull NoneFeatureConfiguration config,
+                                   @Nonnull LevelHeightAccessor accessor) {
 
             // Turns the chunk coordinates into actual coordinates we can use. (Gets center of that chunk)
-            int x = (chunkX << 4) + 7;
-            int z = (chunkZ << 4) + 7;
+            int x = (chunkPos.x << 4) + 7;
+            int z = (chunkPos.z << 4) + 7;
 
             /*
              * We pass this into addPieces to tell it where to generate the structure.
@@ -108,7 +107,7 @@ public class MondroveBundleStructure extends Structure<NoFeatureConfig> {
              * structure will spawn at terrain height instead. Set that parameter to false to
              * force the structure to spawn at blockpos's Y value instead. You got options here!
              */
-            BlockPos blockpos = new BlockPos(x, 0, z);
+            var blockpos = new BlockPos(x, 0, z);
 
             /*
              * If you are doing Nether structures, you'll probably want to spawn your structure on top of ledges.
@@ -120,9 +119,9 @@ public class MondroveBundleStructure extends Structure<NoFeatureConfig> {
             //IBlockReader blockReader = chunkGenerator.getBaseColumn(blockpos.get(X), blockpos.getZ());
 
             // All a structure has to do is call this method to turn it into a jigsaw based structure!
-            JigsawManager.addPieces(
+            JigsawPlacement.addPieces(
                     dynamicRegistryManager,
-                    new VillageConfig(() -> dynamicRegistryManager.registryOrThrow(Registry.TEMPLATE_POOL_REGISTRY)
+                    new JigsawConfiguration(() -> dynamicRegistryManager.registryOrThrow(Registry.TEMPLATE_POOL_REGISTRY)
                             // The path to the starting Template Pool JSON file to read.
                             //
                             // Note, this is "structure_tutorial:run_down_house/start_pool" which means
@@ -130,22 +129,22 @@ public class MondroveBundleStructure extends Structure<NoFeatureConfig> {
                             // "resources/data/structure_tutorial/worldgen/template_pool/run_down_house/start_pool.json"
                             // This is why your poaol files must be in "dta/<modid>/worldgen/template_pool/<the path to the pool here>"
                             // because the game automatically will check in worldgen/template_pool for the pools.
-                            .get(new ResourceLocation(CrypticCosmos.MOD_ID, "mondrove_bundle/start_pool")),
+                            .get(id("mondrove_bundle/start_pool")),
 
                             // How many pieces outward from center can a recursive jigsaw structure spawn.
                             // Our structure is only 1 piece outward and isn't recursive so any value of 1 or more doesn't change anything.
                             // However, I recommend you keep this a decent value like 10 so people can use datapacks to add additional pieces to your structure easily.
                             // But don't make it too large for recursive structures like villages or you'll crash server due to hundreds of pieces attempting to generate!
                             10),
-                    AbstractVillagePiece::new,
+                    PoolElementStructurePiece::new,
                     chunkGenerator,
-                    templateManagerIn,
+                    structureManager,
                     blockpos, // Position of the structure. Y value is ignored if last parameter is set to true.
-                    this.pieces, // The list that will be populated with the jigsaw pieces after this method.
+                    this, // The list that will be populated with the jigsaw pieces after this method.
                     this.random,
                     false, // Special boundary adjustments for villages. It's... hard to explain. Keep this false and make your pieces not be partially intersecting.
                     // Either not intersecting or fully contained will make children pieces spawn just fine. It's easier that way.
-                    true);  // Place at heightmap (top land). Set this to false for structure to be place at the passed in blockpos's Y value instead.
+                    true, accessor);  // Place at heightmap (top land). Set this to false for structure to be place at the passed in blockpos's Y value instead.
             // Definitely keep this false when placing structures in the nether as otherwise, heightmap placing will put the structure on the Bedrock roof.
 
 
@@ -167,18 +166,17 @@ public class MondroveBundleStructure extends Structure<NoFeatureConfig> {
             // By lifting the house up by 1 and lowering the bounding box, the land at bottom of house will now be
             // flush with the surrounding terrain without blocking off the doorstep.
             this.pieces.forEach(piece -> piece.move(0, 1, 0));
-            this.pieces.forEach(piece -> piece.getBoundingBox().y0 -= 1);
-
+            this.pieces.forEach(piece -> piece.getBoundingBox().move(0, -1, 0));
 
             // Sets the bounds of the structure once you are finished.
-            this.calculateBoundingBox();
+            this.getBoundingBox();
 
             // I use to debug and quickly find out if the structure is spawning or not and where it is.
             // This is returning the coordinates of the center starting piece.
             CrypticCosmos.LOGGER.log(Level.DEBUG, "Mondrove Bundle at " +
-                                                  this.pieces.get(0).getBoundingBox().x0 + " " +
-                                                  this.pieces.get(0).getBoundingBox().y0 + " " +
-                                                  this.pieces.get(0).getBoundingBox().z0);
+                                                  this.pieces.get(0).getBoundingBox().minX() + " " +
+                                                  this.pieces.get(0).getBoundingBox().minY() + " " +
+                                                  this.pieces.get(0).getBoundingBox().minZ());
         }
     }
 
